agentcontrol/requirements.txt
pytest==8.4.2
diff-cover==9.7.1
detect-secrets==1.5.0
pip-audit==2.7.3
PyYAML==6.0.3
